feat: straights
This is kind of a black box way of going about it.
We don't know the straights that are being calculated
Only how many points it has
Ideally, we'd know the pairings
Though, all I want to know if the points

Also, I don't like the nested if statements or two pairs straights
It is still linear ignoring the sort algo.
The nested if makes me like this solution a lot less